Improved initialization of the HTTP connection pool based on Czar config
Fixed incorrect comments. Minor reinforcement in the merging handler
to insure the corresponding job is still alive while mering the next batch
of the partial result rows into the destination table. Minor improvementes
in the parameters of methods reporting the number of rows in a result set.
In the previous version of the code the limit was set to 4 bilion rows which may
be potentially exceeded for the LSST-scale catalogs.

+++ src/qdisp/Executive.cc
@@ -635,7 +635,7 @@ void Executive::_setupLimit() {
     _limitSquashApplies = hasLimit && !(groupBy || orderBy || allChunksRequired);
 }
 
-void Executive::addResultRows(uint32_t rowCount) { _totalResultRows += rowCount; }
+void Executive::addResultRows(int64_t rowCount) { _totalResultRows += rowCount; }
 
 void Executive::checkLimitRowComplete() {
     if (!_limitSquashApplies) return;
